wait: Use sqlcmd to detect when the database is actually serving
We have encountered situations where it is possible for the port to be
open, but the system is not ready to actually serve connections.

This change switches from netcat to using `sqlcmd` to execute an actual
command against the database and quit instead, ensuring that the
database is actually ready to serve connections.