Disable aggressive flushing (ok for testing)
Link: https://github.com/microsoft/mssql-docker/issues/355

Latest sqlsrv builds have become way slower. Looking for issues
the linked one above could be a reason. It seems that in CU6,
in order to guarantee data integrity (only for some OSs) they
enabled a more aggressive/often flushing to disk. And that
killed write performance.

In the same issue there are comments about how to disable such
an aggressive flushing, that is perfectly valid for testing.

Hence, doing it and running some unit tests to see if that
gives us some better execution times...